Races v0.1

+++ ApiServer/DAL/RaceDal.cs
@@ -30,7 +30,7 @@ namespace ApiServer.DAL
                                     UserAdmin = (string)sdr["UserAdmin"],
                                     Name = (string)sdr["Name"],
                                     Route = (string)sdr["Route"],
-                                    Cost = (float)sdr["Cost"],
+                                    Cost = (decimal)sdr["Cost"],
                                     Privacy = (Boolean)sdr["Privacy"],
                                     StartDate = (DateTime)sdr["StartDate"],
                                     Category = (string)sdr["Category"],
